Coerce design-time types where needed in type provider
When running with the latest type provider SDK, these three lines making
use of `this` will throw `System.ArgumentException` stating `Incorrect
instance type Parameter name: obj`.

My current understanding is that this occurs at design time in cases
where the `this` value is an instance of a type in the
`ProvidedAssembly`, but the field/method being referenced is part of the
provided type before it was added to the assembly. These fields and
methods could be accessed on the more permanent type via reflection, but
it is easier to just coerce the `this` value. At compile time, this
coercion will not matter, because both the type and the field/method
will be part of the compiled assembly.